+++ src/main/GameStart.java
@@ -41,7 +41,7 @@ public class GameStart extends Application {
         imageView.setPreserveRatio(true);
 
         int startingMoney = Player.getMoney();
-        int startingHealth = Base.getHealth();
+        double startingHealth = Base.getHealth();
 
         String moneyStr = "MONEY: " + String.valueOf(startingMoney);
         String healthStr = "HEALTH: " + String.valueOf(startingHealth) + "hp";
