+++ src/main/PlaceTowers.java
@@ -44,7 +44,7 @@ public class PlaceTowers extends Application {
         imageView.setPreserveRatio(true);
 
         int startingMoney = Player.getMoney();
-        int startingHealth = Base.getHealth();
+        double startingHealth = Base.getHealth();
 
         String moneyStr = "MONEY: " + String.valueOf(startingMoney);
         String healthStr = "HEALTH: " + String.valueOf(startingHealth) + "hp";
